refactor(utils): split token_tracker into package
